+++ .agentplane/tasks/202606020927-HN8GTQ/pr/diffstat.txt
@@ -1,0 +1,4 @@
+ .../src/cli/run-cli.core.pr-flow.test.ts           | 667 -----------------
+ .../run-cli.core.pr-flow.worktree-runtime.test.ts  | 790 +++++++++++++++++++++
+ scripts/oversized-test-baseline.json               |   4 -
+ 3 files changed, 790 insertions(+), 671 deletions(-)
